+++ tests/ecmascript/test-bi-error-instance-custom.js
@@ -39,6 +39,8 @@ all properties (including non-enumerable and inherited)
 "hasOwnProperty"
 "isPrototypeOf"
 "propertyIsEnumerable"
+"__defineGetter__"
+"__defineSetter__"
 ---
 function function false true
 function function false true